Fetch download count on regular basis
In order to keep track of the total downloads we have in Intercity, we created a
bash script on the Intercity.io server that will write to a HTML file how many
downloads per day there have been.

Since those stats are log rotated, we need to fetch them on a regular basis,
that way we have historical data.

All stuff is openly visible for now, only the sidekiq web view is locked behind
a password.